requirements.txt
flask
gunicorn
wemake-python-styleguide
pytest
requests
blinker
flake8
python-dotenv